Create main.c
func_find_srt_file contains a function that returns a string with
the name of the subtitle that is inside the directory where we run the program.
So far we assume that that directory contains just one (or zero) .srt file.

Inside /Functions is every function used inside main.

my_sub_lib.a is a library with my functions. We need to use it to compile
main.c.